pyproject.toml
[project]
name = "tableturn-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogoogle>=5.15.0",
    "aiohttp>=3.12.15",
    "dependency-injector>=4.48.1",
    "fastapi>=0.116.1",
    "fastapi-restful>=0.6.0",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "tortoise-orm[asyncpg]>=0.25.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
